+++ components/SignupForm/SignupForm.tsx
@@ -1,15 +1,10 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import CheckBox from './CheckBox';
 import FormField from '../compound/form/FormField';
 import Button from '../common/Button';
 import Modal from '../common/Modal';
 import useSignupForm from './useSignupForm';
-import Open from '@/public/icons/openEye.svg';
-import Close from '@/public/icons/closeEye.svg';
-import { setItem } from '@/utils/localstorage';
-import login from '@/components/SignupForm/login';
 import PasswordToggleBtn from './PasswordToggleBtn';
 
 const INPUT_FIELD_LIST = [
@@ -36,7 +31,6 @@ const INPUT_FIELD_LIST = [
 ] as const;
 
 export default function SignupForm() {
-  const router = useRouter();
   const {
     formData,
     handleFormChange,
@@ -52,16 +46,6 @@ export default function SignupForm() {
     isPending,
   } = useSignupForm();
 
-  const handleCloseModal = async () => {
-    const credentials = state?.credentials;
-    if (!credentials) return;
-    const res = await login(credentials);
-    if (!res.success) return;
-    router.push('/mydashboard');
-    setItem('userInfo', res.data.user);
-    setItem('accessToken', res.data.accessToken);
-  };
-
   return (
     <form onSubmit={handleFormSubmit} className="flex w-full flex-col gap-6">
       <Modal
@@ -69,7 +53,7 @@ export default function SignupForm() {
         padding="64/40"
         borderRadius="16"
         submitMessage="확인"
-        onClose={handleCloseModal}
+        onClose={onClose}
       >
         {state?.message}
       </Modal>
